// src/components/ShinyText/ShinyText.d.ts
interface ShinyTextProps {
  text: string;
  disabled?: boolean;
  speed?: number;
  className?: string;
  color?: string;
  shineColor?: string;
  spread?: number;
  yoyo?: boolean;
  pauseOnHover?: boolean;
  direction?: 'left' | 'right';
  delay?: number;
}

declare const ShinyText: React.FC<ShinyTextProps>;
export default ShinyText;